+++ next-example/app/page.tsx
@@ -1,10 +1,9 @@
 
-import Image from "next/image";
-import { useCount } from "./useCount";
+import { countStore } from "./useCount";
 import { Counter } from './Counter'
 
 export default async function Home() {
-  const [{ count },] = useCount()
+  const [{ count },] = countStore.use()
   return (
     <>
       <div className="text-xl" test-id="server-count">{count}</div>