+++ src/create.ts
@@ -105,3 +105,5 @@ export function create<T>(
   };
   return [useOne, store] as const;
 }
+
+export type Store = ReturnType<typeof create>[1];