+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # CHANGELOG
 
+### 1.3.0-beta.0
+
+- Feat: add persist store, support react-native and web
+
+---
+
+Usage:
+
+```ts
+import { create } from 'use-one';
+import { persistStore, wrapState, isClient } from 'use-one/persist';
+
+const initialState = wrapState({ count: 0 }); // -> { ready: false, count: 0 }
+const [use, store] = create(initialState);
+
+console.log('isClient', isClient);
+isClient && persistStore(store, { key: '@CACHE_KEY', debounce: 100 });
+```
+
 ### 1.2.0-beta.0
 
 - Feat: support server components (Inspired from https://www.npmjs.com/package/server-only?activeTab=code)